Video with controls js
Now the nice coloring for the videos works on click for playing. It should also be less jank in general since this fixes the issue of it only reacting to clicks rather than if the video is actually playing. So, now the video sound play on a click or with the controls and the border color should work as well.